refactor(room): extract the shared participant base
The private-message inbox and the leave lifecycle were implemented
twice — ServerLocalParticipant and ClientParticipantBase carried
verbatim copies of listen/_deliverMessage/onLeave/_onLeft/_assertActive
(and had already diverged once, by the suppression line). They now
share ParticipantBase in shared.ts, next to RoomState which set the
pattern; flavors keep only their transport and error pipeline.

Also folds two smaller both-sides seams into the shared module:
- normalizeJoinOptions() is used by ClientRoom.join() too instead of
  inline re-validation
- RoomState takes `closed` as a constructor option instead of having
  the client poke the field after construction
